Add files via upload
For my first design I wanted a simple home page that showed all the information clearly and neatly. Because my graphic design is so different from my illustrations I wanted to separate them and put one image to represent them. I wanted to use a font that had a fantasy feel to it because that's how a lot of my art feels.