Add files via upload
For my second design I wanted to highlight one of my illustrations and use it as the main image for the home page. I noticed a lot of artist websites used pieces of their art for the home page like Alex Fisher. I also wanted this design to be more colorful and have a more interesting layout for the navigation. 
https://www.alexfisherdesign.ca/