Add files via upload
For my third design I was inspired by Jennifer Xiao’s artists website. Like Xiao's website I wanted a playful design that was fun to look at and navigate. I wanted a similar color palette to my art so I kept a fall color scheme with oranges and greens. I also plan to have an animation for this design. When someone clicks on the hamburger menu the guy sitting on the menu will fall. 
https://www.jenniferxiao.com/comics